Fixed delay being reset on page reload
Slightly hacky solution since SettingsStore makes currentDelay 0 for one render on reload, so this change will set delayState to currentDelay if currentDelay changes while the input isn't selected. This would theoretically break if the user manages to get a frame-perfect click on the delay input after a reload (I can only manage it when running dev server, not with prod).